Fix disk metric bug, add new metrics
There was a bug with disk metrics where the disk.go metrics were not
being added to the go channel.  Metrics are now added to the channel and
exposed in the registry. Added new metrics around duration of each call
the stats engine, as well as the success of each indivial stats engine
called (per key). Changed the label node to be the node Lnn instead of
the node id.  Now label node is ll and label node_id is the id.